sep style sheet for word item and word modal
